Inline framework packages and check for react-dom/server usage
The framework test was changed in gatsby 5.6 via https://github.com/gatsbyjs/gatsby/pull/37508

Therefor framework.test is no longer of type RegExp and fixFrameworkCache would return early. This resulted in the non-creation of framework-[contenthash].js. Instead the framework was then located in app-[contenthash].js

Since I didn't find a good way to get the framework bundles, I just inlined them from the gatsby repos. They were not changed since 2020 (the creation of the framework chunking). So this seems pretty stable.

Further this now checks for react-dom/server and does not include it in the framework bundle as done in the  main gatsby repos by PR linked above.